quantification/requirements.txt
# 核心依赖
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 数据处理
scipy>=1.7.0
scikit-learn>=1.0.0
statsmodels>=0.13.0

# 金融数据
yfinance>=0.1.70
alpha_vantage>=2.3.0
ccxt>=3.0.0

# 数据库
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0

# Web框架
flask>=2.0.0
flask-socketio>=5.0.0
dash>=2.0.0
plotly>=5.0.0

# 配置管理
pyyaml>=6.0
python-dotenv>=0.19.0

# 日期时间
pytz>=2021.3
python-dateutil>=2.8.2

# 异步处理
asyncio>=3.4.3
aiohttp>=3.8.0

# 测试
pytest>=7.0.0
pytest-cov>=3.0.0

# 开发工具
jupyter>=1.0.0
ipython>=8.0.0
black>=22.0.0
flake8>=4.0.0

# 其他
ta-lib>=0.4.24  # 技术分析库
backtrader>=1.9.76.123  # 回测框架
# zipline>=2.0.0  # 另一个回测框架
empyrical>=0.5.0  # 金融绩效指标